requirements/base.txt
# Django base
asgiref==3.5.1
Django==4.0.4
djangorestframework==3.13.1

# Database
# psycopg2==2.9.3
psycopg2-binary==2.9.3
sqlparse==0.4.2
neo4j==4.4.3
neomodel==4.0.8
django-neomodel==0.0.7

# Mongodb
pymongo==4.3.3
dnspython==2.3.0 

# Auth
django-allauth==0.50.0
dj-rest-auth==2.2.4
djangorestframework-simplejwt==5.1.0

# Documentation
drf-yasg==1.20.0

# Utilities
django-extensions==3.1.5
ipython
django-filter==21.1
phonenumbers==8.12.48
openpyxl==3.0.10
hashids==1.3.1
pendulum==2.1.2
django-modeltranslation==0.18.4
ua-parser==0.16.1
django-admin-extra-buttons==1.5.5
django-phonenumber-field==7.0.0
django-positions==0.6.0

# Other Libraries
# fcm-django==1.0.11
hashids==1.3.1
Pillow==9.1.0
django-cors-headers==3.12.0
plivo==4.24.1
django-s3-folder-storage==0.5
pycryptodome==3.15.0  # For blockchain
openai==1.7.1  # LLM

# Celery
celery==5.2.7
django-celery-beat==2.4.0
redis==4.3.1
python-dotenv==0.13.0

# Caching
django-redis==5.2.0

# Monitoring
sentry-sdk==1.14.0
